refactor: Ensure inputs to the loss function have the same size
Somewhere around completing 296 episodes, probably the first time
the loss was being calculated, a warning came up that read- 'Using
a target size (torch.Size([32, 1])) that is different to the input
size (torch.Size([32])). This will likely lead to incorrect results
due to broadcasting. Please ensure they have the same size.'

This commit addresses the warning by changing the target column
vector to a row vector and reducing a dimension