refactor: Remove sampled elements from replay buffer
Earlier sampled elements were not being removed from the replay buffer.
The buffer got updated by removing older values when newer values were
inserted in, owing to the maxlen argument of the buffer(deque). This was
erroneous because of two reasons:

1. The elements removed were the oldest ones which may not necessarily be
the ones that were previously sampled.

2. At each training step, batch-size number of elements were being sampled
from the buffer but just one new element was being added.

Both of the above reasons increased the possibility of reintroducing the
same tuples to the network which it may have already seen, defeating the
purpose of the replay buffer. This commit fixes this error by removing the
sampled elements from the buffer post sampling.